feat: add support for hooks
Any file can now be set as a hook to be executed BEFORE or AFTER applying the profiles. The hook must be a valid shell script.

Examples:
```shell
gitctx file add prescript.sh --hook pre-apply
gitctx file add postscript.sh --hook post-apply
```

When switching profiles, the hooks will be executed before or after applying the profile. Some ideas for this might be a script that cleans some files, or a script that edits permissions of invokes "ssh-add" after switching profiles.